# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Прозрачная Цена API
  version: 1.1.0-rc.1
servers:
  - url: /api
paths:
...
  /logistics/shipments:
    get:
      security:
        - bearerAuth: []
      summary: List shipments
      responses:
        '200':
          description: Shipment list
    post:
      security:
        - bearerAuth: []
      summary: Create shipment
      responses:
        '201':
          description: Shipment created
  /logistics/shipments/{shipmentId}/transition:
    post:
      security:
        - bearerAuth: []
      summary: Transition shipment state
      parameters:
        - in: path
          name: shipmentId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Shipment updated
  /documents:
    get:
      security:
        - bearerAuth: []
      summary: List documents
      responses:
        '200':
          description: Documents list
  /documents/upload:
    post:
      security:
        - bearerAuth: []
      summary: Upload evidence document
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, type]
              properties:
                file:
                  type: string
                  format: binary
                type:
                  type: string
                lotId:
                  type: string
                dealId:
                  type: string
                shipmentId:
                  type: string
      responses:
        '201':
          description: Document stored
  /documents/{documentId}/download:
    get:
      security:
        - bearerAuth: []
      summary: Download or get signed URL for document
      parameters:
        - in: path
          name: documentId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Download payload
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...